requirements.txt
streamlit==1.35.0
pandas==2.2.2
requests
python-dotenv
mysql-connector-python
msal
langchain
langchain-community
faiss-cpu==1.7.4
sentence-transformers
huggingface-hub
tabulate